split GpsWatcher into it's own file
