+++ OpenRA.Mods.RA/OpenRA.Mods.RA.csproj
@@ -108,6 +108,7 @@
     <Compile Include="Traits\Render\RenderShroudCircle.cs" />
     <Compile Include="Traits\SupportPowers\ChronoshiftPower.cs" />
     <Compile Include="Traits\SupportPowers\GpsPower.cs" />
+    <Compile Include="Traits\GpsWatcher.cs" />
     <Compile Include="Traits\SupportPowers\ParatroopersPower.cs" />
     <Compile Include="Scripting\Properties\ChronosphereProperties.cs" />
     <Compile Include="Scripting\Properties\ParadropProperties.cs" />
